requirements.txt
pandas>=1.5.0
python-dateutil>=2.8.2
openpyxl>=3.0.9
xlrd>=2.0.1
odfpy>=1.4.1
pyyaml>=6.0
tabulate>=0.9.0
faker>=8.0.0
numpy>=1.21.0
pdfminer.six>=20221105
pytesseract>=0.3.10
Pillow>=10.0.0
pdf2image>=1.16.3
PyPDF2>=3.0.0
python-dotenv>=0.19.0
pytest>=7.0.0
black>=22.0.0
flake8>=4.0.0
